Notifier/requirements.txt
requests
jsonpickle
PyMongo
pika
mysql-connector-python